split qwen36 moe geometry helpers
