+++ crates/runner/src/qwen36_moe_engine.rs
@@ -15,17 +15,17 @@ use std::path::Path;
 use anyhow::{anyhow, Context, Result};
 use gpu_hal::{set_backend, Backend};
 use model_store::BakedStore;
-use qwen36_moe::config::TextConfig;
 
 use crate::qwen36_moe_bake::ensure_qwen36_bake;
 use crate::qwen36_moe_decode::{
     argmax_bf16_logits, host_final_norm_lm_head, run_chained_decode, run_chained_decode_fast,
     run_chained_decode_fast_with_expert_prefetch, sample_bf16_logits, ExpertPrefetchPhase,
-    ExpertRoute, MultiLayerGeom, XorshiftRng,
+    ExpertRoute, XorshiftRng,
 };
 use crate::qwen36_moe_dry_run::{
     print_report, run_qwen36_moe_dry_run, ContextSizeSource, DryRunReport,
 };
+use crate::qwen36_moe_geom::build_multi_layer_geom;
 use crate::qwen36_moe_host::{host_load_bytes, load_lm_head_bf16, lookup_embed_row};
 use crate::qwen36_moe_layers::{load_layer_buffers, Qwen36WeightMode};
 use crate::qwen36_moe_prefetch::handle_moe_expert_prefetch;
@@ -44,7 +44,7 @@ use crate::qwen36_moe_vmm::{
     moe_island_prefetch_ranks_from_env, moe_island_prefetch_transition_min_observations,
     should_use_qwen36_kv_vmm, virtual_kv_stats_for_layers, MoeExpertVmmMode,
 };
-use crate::registry::{Qwen36MoeKernelParams, RegistryEntry};
+use crate::registry::RegistryEntry;
 
 const MIB: f64 = (1024 * 1024) as f64;
 const QWEN36_NUM_SPECULATIVE_TOKENS: usize = 3;
@@ -159,39 +159,6 @@ pub fn run(cli: &crate::Cli, entry: &RegistryEntry, total_vram: u64) -> Result<(
         cli.dump_last_logits,
     )?;
     Ok(())
-}
-
-/// Build the geometry the chained decoder needs from the parsed config +
-/// the registry's per-family params. Mirrors what
-/// `oracle/qwen36_moe_multilayer_oracle.py` puts in `config` and what
-/// `MultiLayerGeom` consumes.
-fn build_multi_layer_geom(
-    text_config: &TextConfig,
-    kernel_params: &Qwen36MoeKernelParams,
-) -> MultiLayerGeom {
-    MultiLayerGeom {
-        hidden: text_config.hidden_size as i32,
-        vocab: text_config.vocab_size as i32,
-        num_layers: text_config.num_hidden_layers as i32,
-        rms_norm_eps: text_config.rms_norm_eps as f32,
-
-        num_attention_heads: text_config.num_attention_heads as i32,
-        num_kv_heads: text_config.num_key_value_heads as i32,
-        head_dim: text_config.head_dim as i32,
-        rotary_dim: text_config.rotary_dim() as i32,
-        rope_theta: text_config.rope_theta() as f32,
-
-        num_k_heads: text_config.linear_num_key_heads as i32,
-        num_v_heads: text_config.linear_num_value_heads as i32,
-        head_k_dim: text_config.linear_key_head_dim as i32,
-        head_v_dim: text_config.linear_value_head_dim as i32,
-        conv_kernel_dim: text_config.linear_conv_kernel_dim as i32,
-
-        num_experts: kernel_params.num_experts as i32,
-        moe_intermediate: kernel_params.moe_intermediate_size as i32,
-        shared_intermediate: kernel_params.shared_expert_intermediate_size as i32,
-        top_k: kernel_params.top_k as i32,
-    }
 }
 
 /// Tokenize the prompt and run the multi-token decode loop end-to-end:
@@ -1344,180 +1311,4 @@ fn decode_first_token(model_dir: &Path, report: &DryRunReport, kv_fp8: bool) -> 
         geom.rms_norm_eps,
     );
     Ok(argmax_bf16_logits(&logits))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{ExpertRoute, MoeIslandPrefetchMode, MoeRouteTelemetry, MoeTransitionPredictor};
-
-    #[test]
-    fn moe_prefetch_mode_env_accepts_disabled_and_previous_token_aliases() {
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(None).unwrap(),
-            MoeIslandPrefetchMode::Disabled
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("disabled")).unwrap(),
-            MoeIslandPrefetchMode::Disabled
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("previous-token")).unwrap(),
-            MoeIslandPrefetchMode::PreviousToken
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("prev-token")).unwrap(),
-            MoeIslandPrefetchMode::PreviousToken
-        );
-    }
-
-    #[test]
-    fn moe_prefetch_mode_env_accepts_resident_only_aliases() {
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("previous-token-resident")).unwrap(),
-            MoeIslandPrefetchMode::PreviousTokenResidentOnly
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("previous_token_resident")).unwrap(),
-            MoeIslandPrefetchMode::PreviousTokenResidentOnly
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("resident-previous-token")).unwrap(),
-            MoeIslandPrefetchMode::PreviousTokenResidentOnly
-        );
-        assert!(MoeIslandPrefetchMode::from_env_value(Some("resident")).is_err());
-    }
-
-    #[test]
-    fn moe_prefetch_mode_env_accepts_transition_aliases() {
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("transition")).unwrap(),
-            MoeIslandPrefetchMode::Transition
-        );
-        assert_eq!(
-            MoeIslandPrefetchMode::from_env_value(Some("transition-weighted")).unwrap(),
-            MoeIslandPrefetchMode::Transition
-        );
-        assert_eq!(MoeIslandPrefetchMode::Transition.as_str(), "transition");
-        assert!(MoeIslandPrefetchMode::Transition.uses_previous_token_routes());
-        assert!(MoeIslandPrefetchMode::Transition.transition_weighted());
-    }
-
-    #[test]
-    fn moe_transition_predictor_waits_for_warmup_and_scores_repeats() {
-        let mut predictor = MoeTransitionPredictor::new(3, 2);
-        let previous_routes = [10, 20, 30];
-        let routes = [
-            ExpertRoute {
-                rank: 0,
-                expert_idx: 20,
-                weight: 0.5,
-            },
-            ExpertRoute {
-                rank: 1,
-                expert_idx: 99,
-                weight: 0.25,
-            },
-        ];
-
-        predictor.update(&routes, &previous_routes);
-        assert!(predictor.candidates(&previous_routes, 2).is_empty());
-
-        predictor.update(&routes, &previous_routes);
-        assert_eq!(predictor.candidates(&previous_routes, 2), vec![20]);
-
-        let later_routes = [
-            ExpertRoute {
-                rank: 0,
-                expert_idx: 10,
-                weight: 0.5,
-            },
-            ExpertRoute {
-                rank: 1,
-                expert_idx: 20,
-                weight: 0.25,
-            },
-        ];
-        predictor.update(&later_routes, &previous_routes);
-        assert_eq!(predictor.candidates(&previous_routes, 2), vec![20, 10]);
-    }
-
-    #[test]
-    fn moe_route_telemetry_records_previous_rank_transition_matrix() {
-        let mut telemetry = MoeRouteTelemetry::new(3);
-        let previous_routes = [7, 11, 13];
-        telemetry.record_route_observation(
-            &ExpertRoute {
-                rank: 0,
-                expert_idx: 11,
-                weight: 0.5,
-            },
-            &previous_routes,
-        );
-        telemetry.record_route_observation(
-            &ExpertRoute {
-                rank: 1,
-                expert_idx: 7,
-                weight: 0.25,
-            },
-            &previous_routes,
-        );
-        telemetry.record_route_observation(
-            &ExpertRoute {
-                rank: 2,
-                expert_idx: 99,
-                weight: 0.125,
-            },
-            &previous_routes,
-        );
-
-        assert_eq!(telemetry.observations_by_rank, vec![1, 1, 1]);
-        assert_eq!(telemetry.repeated_previous_by_rank, vec![1, 1, 0]);
-        assert_eq!(
-            telemetry.repeated_previous_rank_by_current_rank,
-            vec![vec![0, 1, 0], vec![1, 0, 0], vec![0, 0, 0]]
-        );
-        assert_eq!(
-            telemetry
-                .to_json()
-                .get("repeated_previous_rank_by_current_rank")
-                .unwrap(),
-            &serde_json::json!([[0, 1, 0], [1, 0, 0], [0, 0, 0]])
-        );
-        let json = telemetry.to_json();
-        assert_eq!(
-            json.get("repeated_previous_probability_by_current_rank")
-                .unwrap(),
-            &serde_json::json!([1.0, 1.0, 0.0])
-        );
-        assert_eq!(
-            json.get("same_rank_repeat_probability_by_rank").unwrap(),
-            &serde_json::json!([0.0, 0.0, 0.0])
-        );
-        assert_eq!(
-            json.get("repeated_current_by_previous_rank").unwrap(),
-            &serde_json::json!([1, 1, 0])
-        );
-        assert_eq!(
-            json.get("repeated_current_probability_by_previous_rank")
-                .unwrap(),
-            &serde_json::json!([1.0, 1.0, 0.0])
-        );
-        assert_eq!(
-            json.get("best_previous_rank_by_current_rank").unwrap(),
-            &serde_json::json!([1, 0, null])
-        );
-        assert_eq!(
-            json.get("best_current_rank_by_previous_rank").unwrap(),
-            &serde_json::json!([1, 0, null])
-        );
-        assert_eq!(
-            json.get("best_transition").unwrap(),
-            &serde_json::json!({
-                "current_rank": 0,
-                "previous_rank": 1,
-                "count": 1,
-                "probability_by_current_rank": 1.0,
-            })
-        );
-    }
 }
