+++ crates/runner/src/qwen36_moe_geom.rs
@@ -1,0 +1,37 @@
+use qwen36_moe::config::TextConfig;
+
+use crate::qwen36_moe_decode::MultiLayerGeom;
+use crate::registry::Qwen36MoeKernelParams;
+
+/// Build the geometry the chained decoder needs from the parsed config and
+/// the registry's per-family params. Mirrors what
+/// `oracle/qwen36_moe_multilayer_oracle.py` puts in `config` and what
+/// `MultiLayerGeom` consumes.
+pub(crate) fn build_multi_layer_geom(
+    text_config: &TextConfig,
+    kernel_params: &Qwen36MoeKernelParams,
+) -> MultiLayerGeom {
+    MultiLayerGeom {
+        hidden: text_config.hidden_size as i32,
+        vocab: text_config.vocab_size as i32,
+        num_layers: text_config.num_hidden_layers as i32,
+        rms_norm_eps: text_config.rms_norm_eps as f32,
+
+        num_attention_heads: text_config.num_attention_heads as i32,
+        num_kv_heads: text_config.num_key_value_heads as i32,
+        head_dim: text_config.head_dim as i32,
+        rotary_dim: text_config.rotary_dim() as i32,
+        rope_theta: text_config.rope_theta() as f32,
+
+        num_k_heads: text_config.linear_num_key_heads as i32,
+        num_v_heads: text_config.linear_num_value_heads as i32,
+        head_k_dim: text_config.linear_key_head_dim as i32,
+        head_v_dim: text_config.linear_value_head_dim as i32,
+        conv_kernel_dim: text_config.linear_conv_kernel_dim as i32,
+
+        num_experts: kernel_params.num_experts as i32,
+        moe_intermediate: kernel_params.moe_intermediate_size as i32,
+        shared_intermediate: kernel_params.shared_expert_intermediate_size as i32,
+        top_k: kernel_params.top_k as i32,
+    }
+}
